Moved transition checks up to SceneManager
